+++ example/operationsync/src/main/java/org/apache/iotdb/operationsync/OperationSyncEmpty.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.iotdb.doublewrite;
+package org.apache.iotdb.operationsync;
 
 public class OperationSyncEmpty extends OperationSyncUtil {
